Member

See merge request cloud_track/class_02/web_project3/team03/coursemaker-be!19

+++ src/main/java/coursemaker/coursemaker/domain/course/service/CourseServiceImpl.java
@@ -2,23 +2,26 @@ package coursemaker.coursemaker.domain.course.service;
 
 import coursemaker.coursemaker.domain.course.dto.*;
 import coursemaker.coursemaker.domain.course.entity.CourseDestination;
+import coursemaker.coursemaker.domain.course.exception.TravelCourseDuplicatedException;
+import coursemaker.coursemaker.domain.course.exception.TravelCourseNotFoundException;
 import coursemaker.coursemaker.domain.course.repository.CourseDestinationRepository;
-
 import coursemaker.coursemaker.domain.course.entity.TravelCourse;
+import coursemaker.coursemaker.domain.course.service.CourseDestinationService;
 import coursemaker.coursemaker.domain.course.repository.TravelCourseRepository;
-
-import coursemaker.coursemaker.domain.destination.entity.Destination;
-import coursemaker.coursemaker.domain.destination.repository.DestinationRepository;
+import coursemaker.coursemaker.domain.destination.dto.DestinationDto;
+import coursemaker.coursemaker.domain.tag.dto.TagResponseDto;
+import coursemaker.coursemaker.domain.tag.entity.Tag;
+import coursemaker.coursemaker.domain.tag.service.TagService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -26,14 +29,19 @@ import java.util.stream.Collectors;
 @Transactional
 public class CourseServiceImpl implements CourseService{
 
-    @Autowired
     private final CourseDestinationRepository courseDestinationRepository;
 
-    @Autowired
     private  final TravelCourseRepository travelCourseRepository;
+
+    @Lazy
+    private final CourseDestinationService courseDestinationService;
 
     @Override
     public TravelCourse save(AddTravelCourseRequest request) {
+        Optional<TravelCourse> existingCourse = travelCourseRepository.findByTitle(request.getTitle());
+        if (existingCourse.isPresent()) {
+            throw new TravelCourseDuplicatedException("이미 존재하는 코스입니다. ", "코스 이름: " + request.getTitle());
+        }
         return travelCourseRepository.save(request.toEntity());
     }
 
@@ -50,61 +58,63 @@ public class CourseServiceImpl implements CourseService{
     @Override
     public TravelCourse findById(long id) {
         return travelCourseRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Travel course not found with id: " + id));
+                .orElseThrow(() -> new TravelCourseNotFoundException("존재하지 않는 코스입니다. ", "Course ID: " + id));
     }
 
     @Override
     public TravelCourse update(long id, UpdateTravelCourseRequest request) {
         TravelCourse travelCourse = travelCourseRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Travel course not found with id: " + id));
+                .orElseThrow(() -> new TravelCourseNotFoundException("수정할 코스가 존재하지 않습니다. ", "Course ID: " + id));
         travelCourse.update(request.getTitle(), request.getContent(), request.getDuration(), request.getTravelerCount(), request.getTravelType(), request.getPictureLink());
         return travelCourseRepository.save(travelCourse);
     }
 
     @Override
     public void delete(long id) {
+        if (!travelCourseRepository.existsById(id)) {
+            throw new TravelCourseNotFoundException("삭제할 코스가 존재하지 않습니다. " + id, "Course ID: " + id);
+        }
         travelCourseRepository.deleteById(id);
-    }
-
-    @Override
-    public List<CourseDestinationResponse> findAllCourseDestinations() {
-        return courseDestinationRepository.findAll().stream()
-                .map(CourseDestinationResponse::new)
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public CourseDestinationResponse findCourseDestinationById(long id) {
-        CourseDestination courseDestination = courseDestinationRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Course destination not found with id: " + id));
-        return new CourseDestinationResponse(courseDestination);
-    }
-
-    @Override
-    public CourseDestination addCourseDestination(AddCourseDestinationRequest request) {
-        CourseDestination courseDestination = request.toEntity();
-        return courseDestinationRepository.save(courseDestination);
-    }
-
-    @Override
-    public CourseDestination updateCourseDestination(long id, UpdateCourseDestinationRequest request) {
-        CourseDestination courseDestination = courseDestinationRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Course destination not found with id: " + id));
-        courseDestination.update(request.getVisitOrder(), request.getDate(), request.getDestination());
-        return courseDestinationRepository.save(courseDestination);
-    }
-
-    @Override
-    public void deleteCourseDestination(long id) {
-        courseDestinationRepository.deleteById(id);
     }
 
     @Override
     public TravelCourse incrementViews(long id) {
         TravelCourse travelCourse = travelCourseRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Travel course not found with id: " + id));
+                .orElseThrow(() -> new TravelCourseNotFoundException("코스가 존재하지 않습니다.: ", "Course ID: " + id));
         travelCourse.incrementViews();
         return travelCourseRepository.save(travelCourse);
     }
 
+//    @Override
+//    public List<CourseDestinationResponse> findAllCourseDestinations() {
+//        return courseDestinationRepository.findAll().stream()
+//                .map(courseDestinationService::toResponse)
+//                .collect(Collectors.toList());
+//    }
+//
+//    @Override
+//    public CourseDestinationResponse findCourseDestinationById(long id) {
+//        CourseDestination courseDestination = courseDestinationRepository.findById(id)
+//                .orElseThrow(() -> new IllegalArgumentException("Course destination not found with id: " + id));
+//        return courseDestinationService.toResponse(courseDestination);
+//    }
+//
+//    @Override
+//    public CourseDestination addCourseDestination(AddCourseDestinationRequest request) {
+//        CourseDestination courseDestination = request.toEntity();
+//        return courseDestinationRepository.save(courseDestination);
+//    }
+//
+//    @Override
+//    public CourseDestination updateCourseDestination(long id, UpdateCourseDestinationRequest request) {
+//        CourseDestination courseDestination = courseDestinationRepository.findById(id)
+//                .orElseThrow(() -> new IllegalArgumentException("Course destination not found with id: " + id));
+//        courseDestination.update(request.getVisitOrder(), request.getDate(), request.getDestination());
+//        return courseDestinationRepository.save(courseDestination);
+//    }
+//
+//    @Override
+//    public void deleteCourseDestination(long id) {
+//        courseDestinationRepository.deleteById(id);
+//    }
 }
